Pass through the Nova Server Manager and use it.
